moved a couple of methods from pp into la and refactored things a bit. Renamed properties to be label-specific
