+++ tests/la/test_add_label_type.py
@@ -6,23 +6,23 @@ from spatialproteomics.constants import Dims, Layers
 def test_add_label_type(dataset_full):
     ds = dataset_full.la.add_label_type("Cell type 1")
 
-    assert Layers.PROPERTIES in ds
-    assert Layers.PROPERTIES not in dataset_full
+    assert Layers.LA_PROPERTIES in ds
+    assert Layers.LA_PROPERTIES not in dataset_full
     assert Dims.LABELS in ds.coords
     assert Dims.LABELS not in dataset_full.coords
 
-    assert ds[Layers.PROPERTIES].sel({Dims.PROPS: "_name", Dims.LABELS: 1}).values.item() == "Cell type 1"
-    assert ds[Layers.PROPERTIES].sel({Dims.PROPS: "_color", Dims.LABELS: 1}).values.item() == "w"
+    assert ds[Layers.LA_PROPERTIES].sel({Dims.LA_PROPS: "_name", Dims.LABELS: 1}).values.item() == "Cell type 1"
+    assert ds[Layers.LA_PROPERTIES].sel({Dims.LA_PROPS: "_color", Dims.LABELS: 1}).values.item() == "w"
 
     ds = ds.la.add_label_type("Cell type 2", color="k")
 
-    assert Layers.PROPERTIES in ds
-    assert Layers.PROPERTIES not in dataset_full
+    assert Layers.LA_PROPERTIES in ds
+    assert Layers.LA_PROPERTIES not in dataset_full
     assert Dims.LABELS in ds.coords
     assert Dims.LABELS not in dataset_full.coords
 
-    assert ds[Layers.PROPERTIES].sel({Dims.PROPS: "_name", Dims.LABELS: 2}).values.item() == "Cell type 2"
-    assert ds[Layers.PROPERTIES].sel({Dims.PROPS: "_color", Dims.LABELS: 2}).values.item() == "k"
+    assert ds[Layers.LA_PROPERTIES].sel({Dims.LA_PROPS: "_name", Dims.LABELS: 2}).values.item() == "Cell type 2"
+    assert ds[Layers.LA_PROPERTIES].sel({Dims.LA_PROPS: "_color", Dims.LABELS: 2}).values.item() == "k"
 
 
 def test_label_no_segmentation(dataset_segmentation):
@@ -33,8 +33,8 @@ def test_label_no_segmentation(dataset_segmentation):
 def test_add_duplicate_label_type(dataset_full):
     ds = dataset_full.la.add_label_type("Cell type 1")
 
-    assert Layers.PROPERTIES in ds
-    assert Layers.PROPERTIES not in dataset_full
+    assert Layers.LA_PROPERTIES in ds
+    assert Layers.LA_PROPERTIES not in dataset_full
     assert Dims.LABELS in ds.coords
     assert Dims.LABELS not in dataset_full.coords
