+++ tests/la/test_predict_cell_types_argmax.py
@@ -36,23 +36,23 @@ def test_predict_cell_types_argmax_without_overwriting_existing_annotations(data
     )
 
     # adding the labels
-    ds = dataset_full.pp.add_labels_from_dataframe(df)
+    ds = dataset_full.la.add_labels_from_dataframe(df)
 
     # adding a quantification layer
     ds = ds.pp.add_quantification()
 
     # at this point, there should be some assigned and some unassigned cells
     assert np.all(np.unique(ds.coords["labels"].values) == np.array([0, 1]))
-    assert "CT1" in ds[Layers.PROPERTIES].sel(props=Props.NAME).values
-    assert Labels.UNLABELED in ds[Layers.PROPERTIES].sel(props=Props.NAME).values
+    assert "CT1" in ds[Layers.LA_PROPERTIES].sel(la_props=Props.NAME).values
+    assert Labels.UNLABELED in ds[Layers.LA_PROPERTIES].sel(la_props=Props.NAME).values
 
     ct_dict = {"T_CD4": "CD4", "T_CD8": "CD8"}
     ds = ds.la.predict_cell_types_argmax(ct_dict)
 
     # checking that we have all cell types and no unassigned cells
-    assert "CT1" in ds[Layers.PROPERTIES].sel(props=Props.NAME).values
-    assert "T_CD4" in ds[Layers.PROPERTIES].sel(props=Props.NAME).values
-    assert "T_CD8" in ds[Layers.PROPERTIES].sel(props=Props.NAME).values
-    assert Labels.UNLABELED not in ds[Layers.PROPERTIES].sel(props=Props.NAME).values
+    assert "CT1" in ds[Layers.LA_PROPERTIES].sel(la_props=Props.NAME).values
+    assert "T_CD4" in ds[Layers.LA_PROPERTIES].sel(la_props=Props.NAME).values
+    assert "T_CD8" in ds[Layers.LA_PROPERTIES].sel(la_props=Props.NAME).values
+    assert Labels.UNLABELED not in ds[Layers.LA_PROPERTIES].sel(la_props=Props.NAME).values
     # this check implicitly checks if there are unassigned (0) cells left
     assert np.all(np.unique(ds.coords["labels"].values) == np.array([1, 2, 3]))
